+++ src/caretogether-pwa/src/Families/AgeEditor.tsx
@@ -3,7 +3,7 @@ import { useDirectoryModel } from '../Model/DirectoryModel';
 import { useInlineEditor } from '../Hooks/useInlineEditor';
 import { PersonEditorProps } from "./PersonEditorProps";
 import { AgeInYears, ExactAge } from '../GeneratedClient';
-import { AgeText } from '../AgeText';
+import { AgeText } from './AgeText';
 import { format } from 'date-fns';
 import { DatePicker } from '@mui/x-date-pickers';
