+++ src/caretogether-pwa/src/Referrals/Arrangements/ReopenArrangementDialog.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'react-router';
 import { Arrangement, Person } from '../../GeneratedClient';
 import { useDirectoryModel, usePersonLookup } from '../../Model/DirectoryModel';
 import { useReferralsModel } from '../../Model/ReferralsModel';
-import { UpdateDialog } from '../../UpdateDialog';
+import { UpdateDialog } from '../../Generic/UpdateDialog';
 
 interface ReopenArrangementDialogProps {
   referralId: string,
