+++ src/caretogether-pwa/src/Referrals/CloseReferralDialog.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { FormControl, FormControlLabel, FormLabel, Grid, Radio, RadioGroup } from '@mui/material';
 import { ReferralCloseReason } from '../GeneratedClient';
-import { UpdateDialog } from '../UpdateDialog';
+import { UpdateDialog } from '../Generic/UpdateDialog';
 import { DatePicker } from '@mui/x-date-pickers';
 import { useReferralsModel } from '../Model/ReferralsModel';
