+++ src/caretogether-pwa/src/Requirements/MissingArrangementRequirementRow.tsx
@@ -7,7 +7,7 @@ import { useFamilyIdPermissions } from "../Model/SessionModel";
 import { useDialogHandle } from "../Hooks/useDialogHandle";
 import { FamilyName } from "../Families/FamilyName";
 import { PersonName } from "../Families/PersonName";
-import { IconRow } from "../IconRow";
+import { IconRow } from "../Generic/IconRow";
 import { MissingRequirementDialog } from "./MissingRequirementDialog";
 import { RequirementContext } from "./RequirementContext";
