+++ src/caretogether-pwa/src/Shell/SearchBar.tsx
