+++ src/caretogether-pwa/src/Volunteers/ResetFamilyRoleDialog.tsx
@@ -1,6 +1,6 @@
 import { RoleRemovalReason } from '../GeneratedClient';
 import { useVolunteersModel } from '../Model/VolunteersModel';
-import { UpdateDialog } from '../UpdateDialog';
+import { UpdateDialog } from '../Generic/UpdateDialog';
 
 interface ResetFamilyRoleDialogProps {
   volunteerFamilyId: string,
