+++ src/caretogether-pwa/src/Volunteers/ResetIndividualRoleDialog.tsx
@@ -1,6 +1,6 @@
 import { Person, RoleRemovalReason } from '../GeneratedClient';
 import { useVolunteersModel } from '../Model/VolunteersModel';
-import { UpdateDialog } from '../UpdateDialog';
+import { UpdateDialog } from '../Generic/UpdateDialog';
 
 interface ResetIndividualRoleDialogProps {
   volunteerFamilyId: string,
